MetroX_309/ai_railway_project/requirements.txt
pandas
numpy
scikit-learn
supabase
python-dotenv
joblib
